Make component SnippetBlock standalone
