+++ tsd.math.software---mvp/src/app/main/tsd/components/SnippetBlock/SnippetBlock.css
@@ -1,0 +1,13 @@
+/* Copyright (c) 2024 Tobias Briones. All rights reserved. */
+/* SPDX-License-Identifier: AGPL-3.0-or-later */
+/* This file is part of https://github.com/texsydo/texsydo---mvp */
+
+.snippet-block.compact code {
+    font-size: 0.6875rem;
+}
+
+@media (min-width: 1080px) {
+    .snippet-block.compact code {
+        font-size: 0.75rem;
+    }
+}
